Create toggleable panel
