+++ Assets/Scripts/Dialog/DialogGiver.cs
@@ -13,7 +13,7 @@ public class DialogGiver : MonoBehaviour
     var player = other.GetComponent<ThirdPersonMover>();
     if (player != null) {
       transform.LookAt(player.transform);
-      FindObjectOfType<DialogController>().OpenDialog().StartDialog(_dialog);
+      FindObjectOfType<DialogController>().StartDialog(_dialog);
     }
   }
 }
