+++ Assets/Scripts/UI/ToggleablePanel.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a4924b3029874d388589ccef6059f269
+timeCreated: 1621792782
